Rename files to use UpperCamelCaseCluster/Logic for fixed-label-server
